chore: rename Modal → Modal (case change)
